+++ hazel/src/nativeMain/kotlin/hazel/system/extensions.kt
@@ -2,4 +2,4 @@ package hazel.system
 
 expect inline fun breakpoint()
 
-expect inline fun getThreadId(): UInt
+expect inline fun getThreadId(): ULong
